Fix most flake8 errors, and ignore the last one.
I fixed the trivial flake8 errors by reformatting
the code to go along with PEP8.

The error related to "bare except" wasn't trivial
to fix because I don't know what types of Exception
can be thrown by the crashclouseau models. So, I
decided to ignore that error in the .flake8 file
for the sake of fixing the linting.